Database/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.7
jinja2==3.1.2
python-multipart==0.0.6
